+++ src/components/TodoAddform.jsx
@@ -4,37 +4,36 @@ import { __addTodo } from "../redux/modules/todos";
 import styled from "styled-components";
 import uuid from "react-uuid";
 import CusttomButton from "./CusttomButton";
-import useInput from "../hooks/useInput";
+import { useInput } from "../hooks";
+
+// https://tkdodo.eu/blog/practical-react-query
 
 const TodoAddform = () => {
-  // const [title, setTitle] = useState("");
-  // const [content, setContent] = useState("");
+  const [title, setTitle, resetTitle] = useInput("");
+  const [content, setContent, resetContent] = useInput("");
 
-  const { title, setinputTitle, content, setinputContent, onSubmitHandler } =
-    useInput("");
+  const dispatch = useDispatch();
 
-  // const dispatch = useDispatch();
+  const onSubmitHandler = (e) => {
+    e.preventDefault();
+    if (content === "" || title === "") {
+      window.alert("제목및 내용을 입력하세요");
+      return;
+    }
 
-  // const onSubmitHandler = (e) => {
-  //   e.preventDefault();
-  //   if (content === "" || title === "") {
-  //     window.alert("제목및 내용을 입력하세요");
-  //     return;
-  //   }
+    let NewData = {
+      id: uuid(),
+      title,
+      content,
+      isDone: false,
+      displaytoggle: true,
+    };
 
-  //   let NewData = {
-  //     id: uuid(),
-  //     title,
-  //     content,
-  //     isDone: false,
-  //     displaytoggle: true,
-  //   };
+    dispatch(__addTodo(NewData));
 
-  //   dispatch(__addTodo(NewData));
-
-  //   // setTitle("");
-  //   // setContent("");
-  // };
+    resetTitle();
+    resetContent();
+  };
 
   return (
     <TodoAddformWrap>
@@ -44,14 +43,14 @@ const TodoAddform = () => {
           type="text"
           placeholder="제목을 입력하세요"
           value={title}
-          onChange={setinputTitle}
+          onChange={setTitle}
         />
         <TodoInput
           className="contents"
           type="text"
           placeholder="내용을 입력하세요"
           value={content}
-          onChange={setinputContent}
+          onChange={setContent}
         />
         <CusttomButton>추가</CusttomButton>
       </TodoForm>
