+++ src/hooks/index.js
@@ -1,0 +1,3 @@
+import useInput from "./useInput";
+
+export { useInput };
